+++ components/textbox-rtf/TextBox.web.tsx
@@ -1,10 +1,7 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { useTheme } from "@react-navigation/native";
-import Colors from "@/constants/Colors";
-import ReactQuill from "react-quill";
-import "react-quill/dist/quill.snow.css";
-import { ScrollView } from "react-native";
-import AppLayout from "@/layouts/app-layout";
+import { KeyboardAvoidingView, Platform, ScrollView } from "react-native";
+import { RichText, Toolbar, useEditorBridge } from "@10play/tentap-editor";
 
 interface EditTextAreaProps {
   value: string;
@@ -21,20 +18,33 @@ const EditTextAreaComponent: React.FC<EditTextAreaProps> = ({
 }) => {
   const theme = useTheme();
 
+  // Initialize the editor
+  const editor = useEditorBridge({
+    autofocus: true,
+    avoidIosKeyboard: true,
+    initialContent: value,
+    onChange: () => {
+      editor.getHTML().then((html: string) => {
+        setValue(html);
+      });
+    },
+  });
+
   return (
-    <ScrollView style={{ flex: 1 }}>
-      <ReactQuill
-        theme="snow"
-        value={value as string}
-        onChange={setValue}
-        placeholder={placeholder}
-        style={{
-          flexGrow: 1,
-          backgroundColor: Colors(theme).background,
-          marginTop: 10,
-        }}
-      />
+    <ScrollView
+      style={{ flex: 1 }}
+      contentContainerStyle={{
+        flexGrow: 1,
+      }}
+    >
+      <RichText editor={editor} />
+      <KeyboardAvoidingView
+        behavior={Platform.OS === "web" ? "padding" : "height"}
+      >
+        <Toolbar editor={editor} />
+      </KeyboardAvoidingView>
     </ScrollView>
   );
 };
+
 export default EditTextAreaComponent;
